requirements.txt
streamlit 
tensorflow 
numpy 
pillow 
opencv-python
tensorflow
scikit-learn
joblib
flask